Automatic Emissions-API Software Setup
This patch adds a playbook to setup the software from
https://github.com/emissions-api/emissions-api.git

It is running with gunicorn and is started and enabled via a service file.

Also the software is running as the user emissions-api with less priviledges.
Closes #15.

The user emissions-api does own all software specific files, so this is also
preparation for #5

Related to #16